\name{bodyFat}
\alias{bodyFat}
\docType{data}
\title{Data on body fat}
\description{Lists estimates of the percentage of body fat determined by underwater
weighing and various body circumference measurements for 252 men.
}
\usage{data("bodyFat")}
\format{
  A data frame with 252 observations on the following 15 variables. 
  \describe{
    \item{\code{BD}}{Density determined from underwater weighing}
    \item{\code{PBF}}{Percent body fat from Siri's (1956) equation}
    \item{\code{Age}}{Age (years)}
    \item{\code{WGT}}{Weight (lbs)}
    \item{\code{HGT}}{Height (inches)}
    \item{\code{NCK}}{Neck circumference (cm)}
    \item{\code{CST}}{Chest circumference (cm)}
    \item{\code{ABD}}{Abdomen 2 circumference (cm)}
    \item{\code{Hip}}{Hip circumference (cm)}
    \item{\code{TGH}}{Thigh circumference (cm)}
    \item{\code{KN}}{Knee circumference (cm)}
    \item{\code{ANK}}{Ankle circumference (cm)}
    \item{\code{BCP}}{Biceps (extended) circumference (cm)}
    \item{\code{FA}}{Forearm circumference (cm)}
    \item{\code{WRT}}{Wrist circumference (cm)}
  }
}

\source{
  Statlib.

  \url{http://lib.stat.cmu.edu/datasets/bodyfat}
}

\keyword{datasets}
