\name{malp}
\alias{malp}
\title{
Maximum Agreement Linear Predictor
}
\description{
This is the main function to estimate a linear model using the maximum
agreement predictor.
}
\usage{
malp(formula, data)
}
\arguments{
  \item{formula}{A formula for the linear model}
  \item{data}{A \code{data.frame} containing all variables included in
    the \code{formula}}
}


\keyword{agreement, predictor}
